@@ -1,13 +1,10 @@
 name: wave_builder
 description: A utility for building a wave file by concatenating wave files and periods of silence.
-version: 1.0.4
+version: 1.1.4
 homepage: https://github.com/timkellypa/wave_builder
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
-
-#dependencies:
-#  path: ^1.6.0
+  sdk: '>=2.12.1 <3.0.0'
 
 dev_dependencies:
   pedantic: ^1.8.0
